pyproject.toml
[project]
name = "workshop-langgraph-langfuse-101"
version = "0.1.0"
description = "Workshop on LangGraph & Langfuse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-openai>=0.3.17",
    "langfuse>=2.60.5",
    "langgraph>=0.4.5",
    "textual>=0.47.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-perplexity>=0.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
